// src/webparts/STSMarketingCalendar/interfaces/IEvent.ts
export interface IEvent {
  ID: number;
  Title: string;
  NewTitle: string;
  PublishDate: string;
  Requesteddate: string;
  Status: string;
  Quarter: string;
  Promoted_x0020_asset: string; 
  IndustryId: number;
  Sub_x002d_team: string;
  SocialChannels: string;
  Marketinggoals: string;
  CampaignId: number; 
  Solutions: string;    
  Comments: string;
  Created: string; 
}