// src/webparts/STSMarketingCalendar/interfaces/IResponseItem.ts
import { IResponseFile } from "./IResponseFile";
// create PnP JS response interface for Item
export interface IResponseItem {
  ID: number;
  Title: string;
  NewTitle: string;
  PublishDate: string;
  Requesteddate: string;
  Status: string;
  Quarter: string;
  Promoted_x0020_asset: string; 
  IndustryId: number;
  Sub_x002d_team: string;
  SocialChannels: string;
  Marketinggoals: string;
  CampaignId: number; 
  Solutions: string;    
  Comments: string;
  Created: string; 
}